+++ config/frontburner.config.js
@@ -1,4 +1,9 @@
 module.exports = {
+	prefixes: [
+		'//',
+		'/*',
+		'///'
+	],
 	keywords: [
 		'FIXME',
 		'TEMP',